[kafkazk] kafka objects
